AI Agent for Ecommerce/requirements.txt
# Core backend
fastapi
uvicorn
pydantic
sqlite3  

# Data handling
pandas

# AI + LLM
google-generativeai
python-dotenv

# Frontend & visualization
streamlit
matplotlib


